[test] Made test files for all tested classes.
