+++ test/com/Jeans/InsidePartTest.java
@@ -1,0 +1,18 @@
+package com.Jeans;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+
+public class InsidePartTest {
+    @Test
+    public void checkInsidePartConstructor() {
+
+        InsidePart insidePart = new InsidePart("Cotton", 5, "black");
+
+        Assertions.assertEquals("Cotton", insidePart.getMaterial());
+        Assertions.assertEquals(5, insidePart.getMaterialSize());
+        Assertions.assertEquals("black", insidePart.getColor());
+
+    }
+}
