+++ test/com/Jeans/ZipperTest.java
@@ -1,0 +1,19 @@
+package com.Jeans;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+
+public class ZipperTest {
+    @Test
+    public void checkZipperConstructor() {
+
+        Zipper zipper = new Zipper("Steel", 0.1, "black", 0.15);
+
+        Assertions.assertEquals("Steel", zipper.getMaterial());
+        Assertions.assertEquals(0.1, zipper.getMaterialSize());
+        Assertions.assertEquals("black", zipper.getColor());
+        Assertions.assertEquals(0.15, zipper.getZipperLength());
+
+    }
+}
